backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# DB
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Auth (Firebase ID token verification)
pyjwt>=2.8.0
cryptography>=42.0.0
